requirements.txt
Django==1.10
django-pagedown==0.1.1
Markdown==2.6.6
psycopg2==2.6.2
Pygments==2.1.3
